+++ tests/unit/ViewModel/PageLayout/AbstractPageContentViewTest.php
@@ -20,7 +20,7 @@ class AbstractPageContentViewTest extends TestCase
             $subject
         );
         $this->assertInstanceOf(
-            PageContentView::class,
+            ViewModel\NestedChildView::class,
             $subject
         );
         $this->assertInstanceOf(