+++ tests/unit/ViewModel/PageLayout/StaticPageContentViewTest.php
@@ -4,7 +4,9 @@ namespace test\unit\ViewModel\PageLayout;
 
 use Ingenerator\KohanaView\Exception\UnassignedViewVarException;
 use Ingenerator\KohanaView\TemplateSpecifyingViewModel;
+use Ingenerator\KohanaView\ViewModel\NestedChildView;
 use Ingenerator\KohanaView\ViewModel\PageContentView;
+use Ingenerator\KohanaView\ViewModel\PageLayout\AbstractPageContentView;
 use Ingenerator\KohanaView\ViewModel\PageLayout\StaticPageContentView;
 use PHPUnit\Framework\TestCase;
 use test\mock\ViewModel\PageLayout\DummyPageLayoutView;
@@ -15,7 +17,7 @@ class StaticPageContentViewTest extends TestCase
     {
         $subject = $this->newSubject();
         $this->assertInstanceOf(StaticPageContentView::class, $subject);
-        $this->assertInstanceOf(PageContentView::class, $subject);
+        $this->assertInstanceOf(NestedChildView::class, $subject);
     }
 
     public function test_it_is_a_template_specifying_view()